add errorhandler on DeleteEventUseCase

+++ src/modules/events/useCases/deleteEventUseCase/DeleteEventUseCase.ts
@@ -1,4 +1,5 @@
 import { inject, injectable } from "tsyringe";
+import AppError from "../../../../errors/AppError";
 import { IOrganizerRepository } from "../../../organizers/repositories/IOrganizerRepository";
 import { IDeleteEventDTO } from "../../dtos/IDeleteEventDTO";
 import { IEventRepository } from "../../repositories/IEventRepository";
@@ -14,6 +15,8 @@ class DeleteEventUseCase{
 
   async execute({organizerId,eventId}:IDeleteEventDTO):Promise<void>{
     const organizer = await this.organizersRepository.findById(organizerId);
+
+    if(!organizer) throw new AppError("Organizer don't exist",404);
 
     await this.eventsRepository.delete(organizer,eventId);
   }
